fix(SkillBadge): Base for working SkillBadge story
